Change data in database

+++ app/src/main/java/com/example/bionintelligence/domain/entities/CalculateH2OEntity.java
@@ -1,7 +1,7 @@
 package com.example.bionintelligence.domain.entities;
 
 public class CalculateH2OEntity {
-    public double sf_zpv;
+    public int sf_zpv;
     public double waterConsumption_value;
 
 }
